+++ src/constants.ts
@@ -1,2 +1,2 @@
-export const APP_VERSION = "2.0.6";
-export const DEV_VERSION = "v2.0.6";
+export const APP_VERSION = "3.0.0";
+export const DEV_VERSION = "v3.0.0";